Extract SponsorshipTable component and add Collect all button
